Setting Hilt and Refactoring
