+++ .idea/misc.xml
@@ -11,6 +11,7 @@
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/layout/fragment_symbol.xml" value="0.28396739130434784" />
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/layout/item_recycler_park.xml" value="0.18478260869565216" />
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/layout/item_recycler_symbol.xml" value="0.28396739130434784" />
+        <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/layout/item_recycler_symbol_result.xml" value="0.15526849037487336" />
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/layout/symbol_dialog.xml" value="0.26902173913043476" />
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/menu/navi_menu.xml" value="0.29375" />
         <entry key="..\:/Users/snowc/MapleCal/app/src/main/res/menu/option_menu.xml" value="0.29375" />
