+++ app/src/main/java/com/example/maplecal/MapleCalApplication.kt
@@ -1,0 +1,7 @@
+package com.example.maplecal
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MapleCalApplication : Application()
