+++ app/src/main/java/com/example/maplecal/SymbolData.kt
