+++ app/src/main/java/com/example/maplecal/data/GrowthRepository.kt
@@ -1,4 +1,7 @@
 package com.example.maplecal.data
 
-class GrowthRepository {
+class GrowthRepository(
+    private val expLocalDataSource: ExpLocalDataSource,
+    private val expRemoteDataSource: ExpRemoteDataSource
+) {
 }
