+++ app/src/main/java/com/example/maplecal/data/ParkRepository.kt
@@ -1,4 +1,7 @@
 package com.example.maplecal.data
 
-class ParkRepository {
+class ParkRepository(
+    private val parkLocalDataSource: ParkLocalDataSource
+) {
+
 }
