+++ app/src/main/java/com/example/maplecal/domain/GetGrowthUsecase.kt
@@ -1,0 +1,8 @@
+package com.example.maplecal.domain
+
+import com.example.maplecal.data.GrowthRepository
+
+class GetGrowthUsecase(
+    private val growthRepository: GrowthRepository
+) {
+}
