+++ app/src/main/java/com/example/maplecal/hyper/HyperViewModel.kt
@@ -1,4 +1,25 @@
 package com.example.maplecal.hyper
 
+import com.example.maplecal.HyperData
+
 class HyperViewModel {
+    fun getHyper() = arrayOf(
+        HyperData(0, "STR", 0),
+        HyperData(1, "DEX", 0),
+        HyperData(2, "INT", 0),
+        HyperData(3, "LUK", 0),
+        HyperData(4, "HP", 0),
+        HyperData(5, "MP", 0),
+        HyperData(6, "DF/TF/PP", 0),
+        HyperData(7, "크확", 0),
+        HyperData(8, "크뎀", 0),
+        HyperData(9, "방무", 0),
+        HyperData(10, "데미지", 0),
+        HyperData(11, "보공", 0),
+        HyperData(12, "일몹뎀", 0),
+        HyperData(13, "내성", 0),
+        HyperData(14, "공/마", 0),
+        HyperData(15, "경험치", 0),
+        HyperData(16, "포스", 0)
+    )
 }
