+++ app/src/main/java/com/example/maplecal/park/ParkFragment.kt
@@ -10,7 +10,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.maplecal.R
 import com.example.maplecal.ParkData
 import com.example.maplecal.databinding.FragmentParkBinding
-import com.example.maplecal.model.getPark
 
 class ParkFragment : Fragment(), CompoundButton.OnCheckedChangeListener {
     private lateinit var binding: FragmentParkBinding
@@ -31,7 +30,15 @@ class ParkFragment : Fragment(), CompoundButton.OnCheckedChangeListener {
     }
 
     private fun initView(savedInstanceState: Bundle?) {
-        parks = getPark()
+        parks = arrayOf(
+            ParkData(0, R.drawable.sunday, "일요일 훈장", "0", "0"),
+            ParkData(1, R.drawable.monday, "월요일 훈장", "0", "0"),
+            ParkData(2, R.drawable.tuesday, "화요일 훈장", "0", "0"),
+            ParkData(3, R.drawable.wednesday, "수요일 훈장", "0", "0"),
+            ParkData(4, R.drawable.thursday, "목요일 훈장", "0", "0"),
+            ParkData(5, R.drawable.friday, "금요일 훈장", "0", "0"),
+            ParkData(6, R.drawable.saturday, "토요일 훈장", "0", "0")
+        )
         initParkRecyclerView(savedInstanceState)
         initParkCheckBox()
         initButton()
