--- app/src/main/java/com/example/maplecal/symbol/SymbolRecyclerViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.maplecal.symbol
-
-class SymbolRecyclerViewModel {
-}
