+++ app/src/main/java/com/example/maplecal/symbol/SymbolViewModel.kt
@@ -1,4 +1,21 @@
 package com.example.maplecal.symbol
 
-class SymbolViewModel {
+import androidx.lifecycle.ViewModel
+import com.example.maplecal.R
+import com.example.maplecal.SymbolData
+import com.example.maplecal.domain.GetSymbolUsecase
+
+class SymbolViewModel(
+    private val getSymbolUsecase: GetSymbolUsecase
+) : ViewModel() {
+    val symbols = arrayOf(
+        SymbolData(0, R.drawable.longways, "소멸의 여로", "미니게임\n(0 or 1)", "1", "1", "1"),
+        SymbolData(1, R.drawable.chuchu, "츄츄 아일랜드", "미니게임\n(0 or 1)", "1", "1", "1"),
+        SymbolData(2, R.drawable.lehlne, "레헬른", "드브 층수", "1", "1", "0"),
+        SymbolData(3, R.drawable.arcana, "아르카나", "스세 점수", "1", "1", "30000"),
+        SymbolData(4, R.drawable.moras, "모라스", "미니게임\n(0 or 1)", "1", "1", "1"),
+        SymbolData(5, R.drawable.espa, "에스페라", "미니게임\n(0 or 1)", "1", "1", "1"),
+        SymbolData(6, R.drawable.cernium, "세르니움", "세르니움(후)\n(0 or 1)", "1", "1", "1"),
+        SymbolData(7, R.drawable.arx, "아크르스", "", "1", "1", "0")
+    )
 }
